+++ Source/palette.cpp
@@ -41,9 +41,9 @@ void ApplyGamma(SDL_Color *dst, const SDL_Color *src, int n)
 	g = sgOptions.Graphics.nGammaCorrection / 100.0;
 
 	for (i = 0; i < n; i++) {
-		dst[i].r = pow(src[i].r / 256.0, g) * 256.0;
-		dst[i].g = pow(src[i].g / 256.0, g) * 256.0;
-		dst[i].b = pow(src[i].b / 256.0, g) * 256.0;
+		dst[i].r = static_cast<Uint8>(pow(src[i].r / 256.0, g) * 256.0);
+		dst[i].g = static_cast<Uint8>(pow(src[i].g / 256.0, g) * 256.0);
+		dst[i].b = static_cast<Uint8>(pow(src[i].b / 256.0, g) * 256.0);
 	}
 	force_redraw = 255;
 }
